[fdf] Add support for stable bus topology info
This change allows bus drivers to supply bus information when they
create nodes, as well as provide information about the stability of the
bus nodes it provides. Lastly it provides a way for drivers to expose a
token which can be exchanged but the driver's clients to determine what
the bus topology for these drivers should be.

The bus topology can be reasoned about to such that clients to
reasonably identify the devices they are talking to and ensure it won't
change. Because it is structured, clients also will not need to parse
the bus topology with things like a regex to understand the
underlying bus topology for the device.

Lastly, tokens are used to prevent drivers from both learning about the
the bus topology as well as ensure the client can trust the bus
topology.

Change-Id: I4a26e920ebb09236aba63e74a50b0cbe1e0f70f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148812
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>