[archivist] Simplify shutdown logic
I suspect a bug in this logic was the root cause of an old rare flake
in Archivist's Stop integration tests. It could have been the case that
the `logs_repo.wait_for_termination` future was resolving *before* we
got the stop signal which would lead to not ingesting logs/inspect and
that test flaking. I realized this in the follow-up when I was
observing tests consistently failing due to this.

Multiply: archivist_integration_test
Multiply: archivist-tests
Change-Id: I1584cd18688606d899e7c901867ae31a4e998c4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155572
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>