[ffx][virtualization] Fix build
https://fxrev.dev/1153934 made ffx_tool() require an `sdk_target_name`
when `sdk_category` is specified. `ffx_guest_tool` is not in an SDK, so
it does not need either.

https://fxrev.dev/1153935 increased the scope of marker checks, so now
`ffx_guest` needs to specify "internal" to avoid defaulting to
"partner" because it uses `fuchsia.virtualization`, which is not in any
SDK category.

Bug: 314822328
Test: Add "//src/virtualization/ffx:ffx_guest_tool" to `host_labels` and `fx gen`.
Change-Id: I265ea042e348d6da7c6163c66f0e2e1e8fe3b2ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156574
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>