[kernel][arm64] remove tracking of cluster and cpu id separately
Only one piece of code was using the cluster and cpu number tracking and
it really only needed the raw MPIDR which was being tracked.

Remove the extra arrays that track cluster number (AFF1).

Requires rewriting the logic in the GICv3 SGI firing routine, but is the
same efficiency as before.

Change-Id: I7ca21b79496b61d0eaa546ea255f20e33a362a2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153473
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>