[kernel][arm64] hard set SCTLR_EL1 when initializing cpus instead of orring in bits
More and more bits are being added in newer versions of the ARM spec,
but they're all defaulting to a safe state if 0, so instead of orring in
the initial state of the cpu, hard set it to just the bits we need.

Also update the register definition for newer bits in the ARMv9 spec.

Change-Id: Ifb21abbcaf16dae86b386976fc9501a5c7e5e783
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154860
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>