[sandbox][components] support update notifiers being registered on dicts
Adds support for update notifiers to be registered on sandbox
dictionaries. These are functions which will be invoked whenever an
entry is added to or removed from the dictionary on which it's
registered.

Also uses this change to keep VFS directories created from a dictionary
up-to-date when the contents of the dictionary changes.

Change-Id: I0947d81cca58d883611e90e9bb5890df9f08c668
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136134
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>