[self_profile_report] Add custom leaf rollups
The current self profiling report builds a rollup of all leaf durations
and displays a subset of those taking more than a fixed threshold of
the root duration. This is helpful to help identify top CPU contenders
across the entire profiled CUJ.

For the performance analysis of a complex CUJ, such as system booting,
users may wish to see a rollup of the leaf durations with a custom
filter.

This changes adds a `SelfProfilesReport::add_rollup()` API that allows
users to additionally display rollup of leaf durations based on prefixes.

Bug: b/366533140

Change-Id: Ie83df5f51381c45c52a6cc28dcea900c10bcbab1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155217
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Ali Zhang <alizhang@google.com>