[rbe][cxx] Set remote compile timeout 10m
The default timeout is 1 hour, which is way too long to
wait for a timeout for compiling.
Reduce to 10 minutes for compiling, which is already
more generous than it should be.
(Reduce further later, if deemed appropriate.)

Linking is not affected in this change.

Change-Id: I56c7f1f40509c16c97c570ee0fb23a66899af427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156573
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: David Fang <fangism@google.com>
Commit-Queue: David Fang <fangism@google.com>