[fxfs] handle unaligned allocate calls
We need to be able to handle if the offset and/or length passed to
allocate is not block aligned, since there is no requirement that it is
and regardless, crashing like we did before is not the right response.

Posix fallocate stipulates that the filesystem can allocate more than
was asked for to aligned with block sizes, so we round the range
appropriately, taking care to make sure the recorded size is still
correct and testing that the tail block is properly zeroed.

Bug: 293943124
Change-Id: I02923231a9c189ebe23524340a45e514664fe24b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156212
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>