[fdf] Major refactors of driver test realm
This changes the following stuff in the driver test realm:
 - split out the various servers provided in the driver_test_realm.cc
 - added more methods to the internal protocol used by the resolver
 - the resolver is now responsible for compiling the list of drivers
 - the list of drivers is delivered dynamically through fidl
 - removes some unused routes

This work is done in order for us to be able to collapse the topology of
the driver test realm to no longer have two layers of realm builder.
That is done in the next cl on the stack.

Change-Id: Ifd209117e63b0baed765d51932b20163f0484f03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144152
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>