[netstack3] Fix race in no-op NAT configuration
It's currently possible for NAT not to be configured for a connection by
the time it's inserted in the connection tracking table. For example:
 * If no NAT rules were installed before the connection is finalized,
   NAT is skipped entirely and left unconfigured for the connection
 * If the packet had DNAT performed on it, it would no longer be
   recognizable as part of the same connection due to
   https://fxbug.dev/371017876, and therefore not have SNAT configured

This causes a race condition where subsequent packets on the same flow
will attempt to configure no-op NAT for the connection. Typically NAT is
performed on an owned connection before it's inserted in the conntrack
table, but once it's inserted, we are dealing with a shared reference.
This means two packets can race to configure no-op NAT for the same
flow, and the loser currently causes a panic because we expect NAT to be
unset when we configure it.

Change the panic to only occur in the case where we are holding an owned
connection that is not inserted in the table yet.

The test that exercises this race is non-deterministic; it's not written
as a loom test because loom doesn't support mock versions of the types
relevant to the race (like once_cell::sync::OnceCell). But it does fail
reliably within a handful of runs locally without the fix.

Fixed: 377741038
Multiply: netstack3-core-filter-integration-test
Change-Id: I633497e0f9797f025ce0c2f547fbb2e5b29b0edd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155298
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>