[rust][fidl] Bits support for new Rust bindings
This change also refactors the way constant values are emitted to
support bits more comprehensively. One of the compound identifiers
included in the IR turned out to be a more specialized subset of
compound identifiers, and may include a member path as well. This wasn't
as much of a problem prior to bits support.

Change-Id: I11680ffd421770196acbcd44773e974e341d683f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144777
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>