[superproject] Roll third_party/pigweed pw_bluetooth: Fix RFCOMM information length off by one
A misreading of the spec led to confusion on whether credits was counted
as part of information_length. It is _not_, but it is included in the
calculation for max allowed payload.

Fix this and standardize on `information` as the name used in the spec
vs payload.

Also add helper flag for checking whether a frame is UIH or not

Add standalone unit tests for RFCOMM.

Original-Bug: 378692444
Original-Bug: 378691959
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/249092
Original-Revision: 40d4c1261275065d30e77ea68fbea69061b70e78
GitOrigin-RevId: 13b81cd5ce5bf308b04663d15970aee06e6dc367
Roller-URL: https://cr-buildbucket.appspot.com/build/8731226526069024577
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: If6886e4585d91318dd85c497591f487fa0bc2338
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156496