[fbl][canary] Print the observed value during a panic.
Print the observed value of the magic number when a fbl::Canary fails
an ASSERT predicate, instead of allowing the code to re-load the magic
number post-predicate failure for panic printing purposes.

In a perfect world, this should never (ever) matter.  That said, when
your Canary assert predicates are firing, the world is a long way from
being perfect, and it can actually be important that we report the
exact value used in the predicate evaluation in order to help
understand what might have went wrong.  See the attached bug for a
practical example.

Bug: 313083326
Change-Id: Ic70d393def2e97a79a63ba8b9af7f2b9436dbb4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1158041
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>