[factory] Migrate Clone -> Clone2
Node.Clone is being replaced by Node.Clone2 as part of the ongoing io2
migration. The latter method is already supported by all in and out of
tree servers at all API levels.

All usages in src/factory are in tests, so we clone all connections with
the original rights instead of ensuring they are readable. This can be
resolved in the future by re-opening the directory with the dot-path if
required.

Bug: 378924331
Change-Id: I95c4ac432a95bbf1e171a659fbc3ec1d53e90901
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155955
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>