[starnix][mm] Fix MremapParamTest.MayMove_Expansion with alternate_anon_allocs
The functional difference with this CL is that we are now locating the
"growth" region of the remapped memory using this line of code:

```
let growth_start_addr = dst_addr + length_to_move;
```

Previously, we were putting the growth region *after* the entire size of
the remapped region, which caused us to overwrite memory that might have
been mapped for another purpose.

We also remove the "allocate_address_range" step for mapping anonymous
memory because that step is not needed now that Starnix is responsible
for selecting addresses.

This remainder of this change refactors `select_address` to return an
enum with fewer options so that we can remove the `unreachable` assert
in favor of the compiler proving those causes are not needed using the
type system.

Test: MremapParamTest.MayMove_Expansion
Bug: b/310255065
Change-Id: Id46fad04ebc002dca24812e345825f307120c079
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159114
Reviewed-by: Maria Glukhova <mariagl@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>