[licenses] Use URL from README.fuchsia file if it exists.
Currently, the "legacy GN inputs" license collection process
finds license files by collecting GN metadata entries and
README.fuchsia files. In the latter, it only cares about the name
of the project and the location of the license file. It then tries
to construct a URL that points to the license file on
cs.opensource.google/fuchsia/...

This doesn't always work. Some dependencies are prebuilts, or
third_party projects, or private dependencies, so their licenses
live somewhere else. We need to also pull the license URL
(if it is provided) from the README.fuchsia file.

We should also do this for gn metadata, but that will come later.

Bug: 375001516
Change-Id: Ib64f91d17e6a157bedb04e0d81e974c8dd4a7125
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1157162
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>