[kernel] s/__code_start/__executable_start where appropriate
This change goes throughout kernel source and replaces the usages of
__code_start that represent "start of the executable" to
__executable_start: these symbols coincide when the kernel is linked
with kernel.ld and loaded as a ZBI, but differ in the elf-kernel.ld +
ELF loading context, where __code_start actually refers to the beginning
of the .text section. This change should be a no-op, and ultimately fix
the main bug preventing loading the arm kernel as an ELF (this
discrepancy causing pathologies at runtime).

Bug: 42164859
Change-Id: I8e2be25a5dfccdc8c8458c8ef532f3165c021007
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154774
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>