[dl] Introduce AdjustLoadBiasAdaptor
This is a wrapper around the ld::ModuleMemory object to be used by
the `DecodeModule` function in the DlImplTests fixture.

AdjustLoadBiasAdaptor will attempt to perform a read on the underlying
memory object using the initial `ptr` value passed in: if that read
fails, it will try to read again with `ptr - load_bias`, as is needed
for linux.

Bug: 324136831
Change-Id: I27ec9310701c82c1d85f7aa7cb4618795ac67a81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147858
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>