[botanist][build] Add virtual_device_spec to environment.
In order for botanist to use ffx emu with the product bundle, it needs
to know which virtual_device spec to run with that corresponds to the
desired emu_dimensions. This change also modifies the existing device
specs to match the 1cpu environment and infra default values. Botanist
will only switch to using the product bundle if the recipes pass the
virtual device spec through to botanist.

Bug: 377332712
Change-Id: Ic345f9b81ea08de1935355686ab11182f9e61f9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153900
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>