[tee] Implement binder calls to connect to TA and invoke a command
This adds binder protocols for connecting to a TA and setting up an RPC
session bound to a session with that TA. virtio-socket ports are
allocated sequentially based on the order they appear in the
configuration exposed by the TA manager. In this change communication
with the TAs themselves is stubbed out.

Bug: 369435662
Change-Id: Iee3b676e51913fc773cb5ba664469f150cecee1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129756
Reviewed-by: Eric Stone <ecstone@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>