[rust] Migrate off of async-io
This migrates our async framework off of async-io
and onto Tokio. Unfortunately, Tokio timers don't work
the same way as Fuchsia timers in the sense that an
executor is required to create them on host,
but on target you don't need an executor to create a timer.
To workaround this limitation, the timer is lazily initialized.

Bug: 378000076
Change-Id: If6c3939930b85ce0b846bea663424b5bb4319731
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159422
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>