[testrunner] Scale test timeouts by a factor determined by an env var.
This will be passed as a flag to botanist which will set it as an env
var in the testrunner env. It will be used on bots that use TCG which
runs much slower than other envs. fxrev.dev/1137346 will set the flag
and both need to go in for this change to take effect.

Bug: 366233201
Change-Id: I77010721927000c3f99ff75516c7385447e6e587
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126631
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>