[component_manager] Support tracing when nested
At startup, nested component_manager will connect to a trace_provider in
its namespace, if it has one. As a proof of concept, the
fuchsia.component/Controller.Start method is now traced. You should be
able to get this trace if you run `component_lifecycle_benchmark-test`
with `ffx trace`, for example.

To support this, component_manager has been branched into two variants,
`//src/sys/component_manager:bin` and `:bin_eng`. The only difference is
that `bin_eng` links in libtrace_provider and enables tracing. Splitting
the targets in this way was necessary because linking libtrace_provider
adds ~1.5M to the binary (due to linking shared libs, it seems like),
which is not a tolerable cost in production. For now, `bin_eng` is only
used in nested component managers.

In the future, we may use `:bin_eng` in eng builds, and support using a
trace provider from a child (it won't be available in the root component
manager's namespace). This will require some assembly plumbing.

Change-Id: Ib1a7fda79379612a664fd4313eb02cb2a23debbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147614
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Gary Bressler <geb@google.com>