[kernel][phys] Drop NVRAM from hand-off RAM description
An NVRAM range should not be treated as RAM, and longer-term we will
make that distinction more official. In the meantime, we begin to not
treat it as such by dropping it from the hand-off RAM description, in
turn dropping it from tracking within a PMM arena. One consequence of
this is the need to update the logic around protecting the physmap
regions corresponding to gaps outside of PMM arenas to explicitly avoid
any NVRAM region.

Bug: 42164859
Change-Id: Ibd671b6f96b55e17ee98b3b12229a75ba609191a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159514
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>