[kernel][phys] Disentangle use_elf_kernel from kernel.phys.next
This change removes the connection between the value of use_elf_kernel
and what value of kernel.phys.next is set.

The way the build integration is currently organized, setting
use_elf_kernel = true results in setting an additional
kernel.phys.next option. We remove this connection and instead have each
of the ZBI and ELF flavors of kernel_elf_binary() override the
distribution_entries to ensure they get packaged as "physboot", which is
the default expected value of kernel.phys.next. This sets the stage for
when we'll eventually delete the ZBI one and rename the ELF one as just
"physboot". Moreover, this conveniently sidesteps an internal hoop that
we'd have to jump through when throwing the use_elf_kernel = true switch
for arm relating to internal CMDLINE-diffing logic around arm products.

Bug: 42164859
Change-Id: Idf3960c8b95cef9bec889cd2ef36e7e9fc812d6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1157393
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>