[archivist] Set up separate loop for handling sockets
This sets up a separate thread for processing log sockets which reduces
the per-socket memory overhead and improves performance.

This also allows us to introduce a delay which prevents Archivist from
waking immediately after every message.

As a result of this change, we can reduce the number of threads used for
the executor down to 1 since the executor is no longer used when reading
sockets.

Change-Id: I0e906cef781edcf5121b873e634209a50bbed5a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149015
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>