[network/tests] Deflake neigh_clear_entries
The test is flaky because during the first round of UDP exchange,
Alice may send multiple NDP solicitations, and Bob will respond each
of them. The race happens between the arrival of Bob's advertisements
and the removal of Alice's NDP entry. If one of the responses arrives
later than the removal, then it will install a new entry, causing the
test to fail.

This change relaxes the test so that when it detects a race, it will
retry up to 5 times.

Bug: 379379169
Multiply: netstack-neighbor-fidl-test
Change-Id: Icacb0ae45cd5b3d5494250eb6bd5e12822efee35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159912
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>