[display][api-types-cpp] Rework AlphaMode.
This CL updates AlphaMode to be its own type (as opposed to an alias for
a FIDL type), so driver code is guaranteed to use IsValid() and ToFidl()
/ ToBanjo(). The update gives AlphaMode constexpr conversion methods, so
it can be used in higher-level types that offer constexpr conversion
to/from FIDL/Banjo.

This CL also calls out that we haven't updated the AlphaMode interface
for API stabilization, so it's not suitable for OOT drivers.

Bug: 42079190
Test: fx test //src/graphics/display
Change-Id: I90461c3e7583e398dfc11f15f6797d2eb5d80d23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159333
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>