[memory] Remove last use of HLCPP in pressure_signaler
This commit remove the last remaining uses of legacy HLCPP bindings in
pressure_signaler that it inherited from memory_monitor; it also
removes pressure_signaler from the visibility allowlist that let it
use HLCPP bindings to begin with. This is a bit of cleanup that should
not impact the behavior of pressure_signaler, and that should make the
code a bit more readable.

Change-Id: Id725ac76f886e20e2c2cc84aab3e569bfb9b1198
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159092
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Pierre Labatut <plabatut@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>