[starnix][cgroup] Build more of cgroup hierarchy
Created `CgroupRoot` for the root of a cgroup hierarchy, as it has
different properties than the rest of the tree. Both `CgroupRoot` and
`Cgroup` implements `CgroupOps`. Their main difference is that `Cgroup`
owns the `FsNode` of its `CgroupDirectoryHandle`, whereas the
`FileSystem` owns the `FsNode` of the `CgroupRoot`.

Each `CgroupDirectoryHandle` has a `Weak<dyn CgroupOps>` to get back the
associated `Cgroup`, or `CgroupRoot`, which it can use to implement its
interface files.

`mknod` is no longer used to create the initial `cgroup.procs` file.
Instead, `cgroup.procs` and `cgroup.controllers` files are created
alongside the directory, and can be seen with `ls` without first being
written to.

Cgroup filesystems are now Uncached, because cgroup interface files are
dynamically generated when the directory is created, and are removed
together with the directory. Additionally they do not emit any inotify
events on Linux. As such, we should implement the cgroup filesystem as
Uncached, and fill in the logic for `create_node_ops` and `lookup`.

Bug: 377712920, 375677856, 373897750
Tested: syscalls test
Change-Id: I657e2acfcd4be5e1694ace158299f26256e7acdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146230
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: David Song <wintermelons@google.com>