[kernel] Temporarily remove KASan support on arm
The previous support was found to be making heap allocations in a
pre-heap context*, and was the source of a mystery failure in a
seemingly unrelated and innocuous VM initialization refactor. Given that
and that this support will need to be holistically redesigned to work in
the physboot imminently, the maintainer found it prudent for use to
remove the existing support altogether for now (on the condition of
figuring out and committing to a timeline for the redesign, which is in
the works now).

*Unknown how this was working! The heap allocations were being made
implicitly in arch_asan_early_init() in the usage of CreateSubVmar() and
Map() in asan_map_shadow_for().

Bug: 379891035
Change-Id: I1db9e1f40c6da99a0dfa78b2ca9f04a800e95c03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160133
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>