[binder-proxy] Add TrustedApp aidl interface and conversion
This adds aidl definitions for the binder TrustedApp protocol and adds
conversion routines to the corresponding FIDL types used to communicate
with the TA manager and application.

Parameters sent over Binder RPC are always copied including any data
contained in buffers. Parameters are only included in the Binder RPC
response when the direction is marked INOUT or OUTPUT. The caller can
retain a copy of any INPUT parameters if necessary.

Change-Id: I09bc9b0e39d0462280606ba2fe34e0460ed7db25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155292
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: James Robinson <jamesr@google.com>