[ffx] ffx --strict target show
Add --strict support to `ffx target show`, which basically means
replacing the `TargetProxy` with a query to the DirectConnector. I chose
to expose the connector by adding a new FfxTool field that provides
the direct connection if one is available, as well as a
`Deferred<TargetProxy>`. If we have a direct connection, use it to
gather connection information; otherwise use the `TargetProxy`.
(Some of the information was also available directly from RCS, so
the code also is changes to use RCS where possible.)

Testing this required exposing a `MockDirectConnector`.

Change-Id: I5768290eb602eb78e37f1fb494a49da35bb05275
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160472
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>