[thread_sampler] Disable Ints While Writing Samples
Interrupts are enabled while we are taking samples because we need to
allow interrupts to usercopy. However, because they aren't disabled,
this allows the possibility that we begin writing, set the pending
write, then are context switched out. Another thread would then attempt
to write a sample on the same cpu, violating the assertion that there
isn't a pending write.

Since the trace buffer is mapped and pinned, it is safe to write to it
while interrupts are off, so we now do so.

Testing:

Without the interrupts change, ThreadSampler.HighFrequency consistently
hits the assertion fail. With the added disabled interrupts it now
passes.

Fixed: 379166110
Change-Id: I23b53278c04e291b04c8b4b64c6743ae2c01f49c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159872
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>