[dl] Libdl Supports Startup Modules
This CL will pass a container of ld::abi::Abi<>::Module data structures
to the RuntimeDynamicLinker ctor: a permanent RuntimeModule data
structure is created for each ABI module and added to the modules_ list.

These modules cannot be unloaded for the lifetime of the program.
Appropriate symbol visibility is assumed to already be set on the abi
module when it is passed to the RuntimeDynamicLinker.

Fixed: 324136831
Change-Id: I0c1d4a4e92780eee110a07fdbb081ee6cb944495
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148633
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>