[console-launcher] Remove use of Clone1
Node.Clone is being replaced by Node.Clone2 as part of the ongoing io2
migration. The latter method is already supported by all in and out of
tree servers at all API levels.

Here we cannot use Clone2 as it does not support sending an event when
the server handles the method. Instead we call Open3 again on the dot
path. Alternatively we could use another blocking method to synchronize
in this case.

Bug: 378924331
Change-Id: I8910f1280686f75f22a40f12721bbbdcf3c4e7e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156196
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>