[usb][adb] Fix race in usb-adb-test
Currently, the test teardown logic calls the UnbindOp() on a background
thread, and then proceeds to tear down the fake root device - which
races with the background thread. If the dtor wins the race, it will
cause a failure because the UnbindOp() was called but never (not
yet) replied to. All we need to do is synchronize the UnbindOp() running
on a background thread to the testing thread via
WaitUntilUnbindReplyCalled(). This blocks TearDown() until the driver's
unbind logic actually replies to the unbind request.

Bug: b/380480663
Test: fx test usb-adb-unittest
Multiply: usb-adb-test
Change-Id: I5d68b181b99f2828be1209b79d4bbdbe4f3f35a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162295
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>