[dl] Use AllocChecker in PopulateStartupModules
Establish an API contract with the caller of PopulateStartupModules, to
pass in and check a fbl::AllocChecker to verify the success/failure of
this function.

PopulateStartupModules arms the function-level alloc-checker with the
overall failure or success of the function (using local alloc-checker
instances for actual allocations).

RuntimeDynamicLinker::Create will use its function-level alloc-checker
to perform the allocation of the RuntimeDynamicLinker and pass into
PopulateStartupModules.

Bug: 324136831
Change-Id: I0aaeda540a1ef02b756d949fb7042553e1fe677d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160474
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>