[component_debug] fdio open_deprecated -> open
The test utils requires the handling of `RealmQueryRequest::Open` which
takes the io1` OpenFlags`. It calls FDIO namespace open method which we
want to transition to call the new Open3 and thus take the io2 `Flags`.

As part of this change, do a mapping from the (io1) OpenFlags to (io2)
Flags. The mapping assumes that the incoming open flags only has
RIGHT_READABLE, RIGHT_WRITABLE, or RIGHT_EXECUTABLE set.

Bug: 376575307
Change-Id: I8dd413fd454342c16ad34efcb752da471fdf4d4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162093
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>