[magma][platform] Create PlatformDeviceDfv2
This is a DFv2 version of the platform device which makes FIDL
calls instead of banjo.

Use this in the msd-arm-mali driver first. It will also be used
in the msd-vsi-vip once it migrates to DFv2.

This should not have any functional changes to the msd-arm-mali driver.

Bug: b/301005872
Change-Id: I21c70fa4ae72bcefd834fe21cf1e6f0d0334692d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1158115
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>