[fxfs] Take a read lock when creating prefetch keys future.
Takes a read lock on the object while we know it still exists and pass
that read lock into the prefetch future so that the object cannot be
deleted in the interim.

Also deleting the `pre_fetch()` wrapper method for keys. As it is not a
safe fire-and-forget operation given the desire for looking at
non-existent keys to be considered a failure of the system. Moves the
logic to make it safe up to `pre_fetch_keys` so that has the only notion
of prefetching left in the system.

Bug: 378508610
Change-Id: Id62360d9fae097f433ed56277609073037f86307
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160854
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>