[devicetree] Fix error check when visit fails
The devicetree::Walk does not end when we return false. It only stops
walking the subtree from the node we returned false. This essentially
overrides the status and therefore the error is not propagated up to the
board driver. This CL caches any error code that occurs during the walk
and propagates it to the board driver.

Test: Introduced an error and ran manager-test to confirm failure.
Change-Id: I51b435eabba19340c4c9a0d1dc4edc866e537776
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1161917
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>