[debugger] Deprecate NotifyComponentDiscovered
The underlying ComponentManager event was deprecated some months ago
(see the bug).

This ended up breaking recursive filters such that tests that have a
programless root component, because we never get the notification for
the matching component URL that is installed by fx test. This manifests
in tests running with `fx test --break-on-failure ...` not breaking at
all on any kind of failure or crash.

The recursive filter installation now happens in response to
ComponentStarting events, but this isn't enough to completely solve the
problem, because these events are not emitted for programless components
that are started lazily. TestManager is changed in
https://fxrev.dev/c/1163032/ to start the test realm's root component
eagerly so that we are guaranteed to get this notification for tests.

Fixed: https://fxbug.dev/375663708
Change-Id: I6112b787d8d59ca93e5d42ffc3e570d89f2e82f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153733
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>