[netstack3] Decrement count when connection removed
Currently, the num_connections field tracking the total number of
connections present in the conntrack table is not decremented when a
connection is removed due to Connection::update returning
ConnectionUpdateAction::RemoveEntry, which happens (for example) when a
TCP connection closes.

This means that even when the table is not full, the num_connections
field can grow over time until netstack *thinks* it is full and starts
incurring thrash by removing existing connections in order to insert new
ones. Fix this by decrementing the counter if a connection is removed.

Fixed: 380887806
Test: netstack3-filter-test
Change-Id: I0b219ef0c72c52e16fc5ce25bf2847616a995a67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163474
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>